Refactoring: b-6-get-movie-by-id 1) Fixed functionality to get full movie info with countries, genres and reviews; 2) Added required methods to service layer and dao layer classes; 3) Removed outdated and added new datasets for tests; 4) Refactored tests accordingly to new changes;

+++ src/main/java/org/nomarch/movieland/dao/jdbc/mapper/CountryRowMapper.java
@@ -9,6 +9,6 @@ import java.sql.SQLException;
 public class CountryRowMapper implements RowMapper<Country> {
     @Override
     public Country mapRow(ResultSet resultSet, int rowNum) throws SQLException {
-        return Country.builder().id(resultSet.getInt("id")).name(resultSet.getString("name")).build();
+        return Country.builder().id(resultSet.getLong("id")).name(resultSet.getString("name")).build();
     }
 }
